api/pyproject.toml
[tool.poetry]
name = "todo"
version = "0.1.0"
description = ""
authors = ["Arash Fatahzade <fatahzade@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.6"
djoser = "^2.0.5"
Django = "^3.1.1"
djangorestframework = "^3.11.1"
django_filter = "^2.3.0"
django-cors-headers = "^3.5.0"


[tool.poetry.dev-dependencies]
flake8 = "^3.8.3"
pudb = "^2019.2"
rope = "^0.17.0"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
